# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pocket Gull Node.js Express Gateway",
    "version": "1.0.0",
    "description": "API Gateway for Pocket Gull - handles secrets, patient state sync, and PubMed proxy services."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    },
    {
      "url": "/api",
      "description": "Relative URL (Proxied via Frontend)"
    }
  ],
  "paths": {
...
    "/api/pubmed/search": {
      "get": {
        "summary": "PubMed Search Proxy",
        "description": "Queries the NCBI Entrez PubMed database for articles using a search term.",
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "required": true,
            "description": "The search query (e.g., 'hypothyroidism symptoms')",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of PubMed search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "esearchresult": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "string"
                        },
                        "retmax": {
                          "type": "string"
                        },
                        "retstart": {
                          "type": "string"
                        },
                        "idlist": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing search term",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Term is required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/pubmed/summary": {
      "get": {
        "summary": "PubMed Summary Proxy",
        "description": "Retrieves the detail summaries for a list of PubMed article IDs.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Comma-separated PubMed IDs (e.g., '32912445,31023902')",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of PubMed article summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "uid": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "pubdate": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          },
                          "authors": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "string"
                                },
                                "authtype": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "ID is required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/orcid/{id}": {
      "get": {
        "summary": "ORCID Profile Proxy",
        "description": "Fetches a public ORCID researcher profile by their ORCID ID (e.g. 0000-0002-1825-0097).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The 16-character ORCID ID",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{4}-\\d{4}-\\d{3}[0-9X]$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of ORCID profile record"
          },
          "400": {
            "description": "Invalid format or missing ID"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/hardware/telemetry": {
      "get": {
        "summary": "Retrieve Node Hardware Telemetry",
        "description": "Gathers live CPU and GPU accelerator telemetry (NVIDIA-SMI, AMD ROCm, Intel XPU, Apple Silicon, and system RAM/load metrics) for dynamic local model routing decisions.",
        "responses": {
          "200": {
            "description": "Successful retrieval of hardware metrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cpuName": { "type": "string", "example": "Intel Core i7-9750H" },
                    "cpuLoadPercent": { "type": "integer", "example": 15 },
                    "systemMemoryTotalGb": { "type": "number", "example": 16 },
                    "systemMemoryUsedGb": { "type": "number", "example": 8.5 },
                    "gpus": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "vendor": { "type": "string", "example": "nvidia" },
                          "name": { "type": "string", "example": "NVIDIA GeForce RTX 2060" },
                          "temperatureC": { "type": "integer", "example": 62 },
                          "utilizationPercent": { "type": "integer", "example": 40 },
                          "memoryTotalMiB": { "type": "integer", "example": 6144 },
                          "memoryUsedMiB": { "type": "integer", "example": 2048 },
                          "memoryFreeMiB": { "type": "integer", "example": 4096 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error gathering metrics"
          }
        }
      }
    },
...
  },
...
}